+++ ChangeLog.md
@@ -5,6 +5,7 @@
 **Released: WiP**
 
 - Added support for loading up the alternates for an article.
+  ([#2](https://github.com/davep/oldas/pull/2))
 
 ## v0.1.0